refactor: KitchenridersClient를 도메인으로 이동
