+++ src/main/java/kitchenpos/order/domain/model/KitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.application;
+package kitchenpos.order.domain.model;
 
 import java.math.BigDecimal;
 import java.util.UUID;
